# KMS.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KigImgSift API",
    "description": "图片分类工具API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:12345",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
    "/api/images": {
      "get": {
        "summary": "获取图片列表",
        "description": "扫描源文件夹并返回所有图片文件的列表",
        "responses": {
          "200": {
            "description": "图片文件列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/image": {
      "get": {
        "summary": "获取图片内容",
        "description": "返回指定路径的图片文件的二进制内容",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "图片文件的相对路径"
          }
        ],
        "responses": {
          "200": {
            "description": "图片文件内容",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "图片文件未找到"
          }
        }
      }
    },
    "/api/move": {
      "post": {
        "summary": "移动图片到分类文件夹",
        "description": "将指定的图片文件移动到对应的分类文件夹中",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["filename", "targetType"],
                "properties": {
                  "filename": {
                    "type": "string",
                    "description": "要移动的图片文件名"
                  },
                  "targetType": {
                    "type": "string",
                    "enum": ["frontal", "side"],
                    "description": "目标分类类型"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件移动成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "源文件不存在"
          },
          "500": {
            "description": "服务器内部错误"
          }
        }
      }
    }
  }
}